Moved Path object and simulate logic out of into.py
