+++ odo/into.py
@@ -1,13 +1,7 @@
 from __future__ import absolute_import, division, print_function
 
-import collections
-import importlib
-import inspect
-import re
-
 from toolz import merge
 from multipledispatch import Dispatcher
-from multipledispatch.dispatcher import str_signature
 from .convert import convert
 from .append import append
 from .resource import resource
@@ -16,145 +10,18 @@ from datashape import discover
 from datashape.dispatch import namespace
 from datashape.predicates import isdimension
 from .compatibility import unicode
+from .simulate import Path
 
 
 if 'into' not in namespace:
     namespace['into'] = Dispatcher('into')
 into = namespace['into']
 
-convert_hop = collections.namedtuple(
-    'ConvertHop', ['source', 'target', 'func', 'mod', 'code'])
-
-class Path(object):
-    """Contains information about the conversion path that odo will take between
-    two types.
-
-    Two types of information are contained - information on an append operation
-    and information on a convert operation. Append function calls will frequently
-    contain convert calls as well, so both the append and convert attributes may
-    have values.
-
-    Positional arguments:
-    source -- The object odo is converting from.
-    target -- The object or type odo is converting to."""
-
-    def __init__(self, source, target):
-        self.source = source
-        self.target = target
-        self.append_sig = None
-        self.append_func = None
-        self.append_mod = None
-        self.append_code = None
-        self.convert_sig = None
-        self.convert_path = []
-
-    def add_append(self, func):
-        """Add information about an append function call to the Path object.
-
-        Positional argument:
-        func -- A function that MultipleDispatch returns for a given append
-            signature."""
-
-        self.append_func = func
-        self.append_mod = inspect.getmodule(func)
-        self.append_code = inspect.getsource(func)
-
-    def add_convert(self, path):
-        """Add information about a convert path to the Path object.
-
-        Positional argument:
-        path -- A list of tuples containing source type, target type, and function
-            that NetworkX returns for a given convert signature.
-
-        Creates convert_hop namedtuples containing source type, target type,
-        function to be called, module that the function exists in, and source
-        code for that function, and appends them to self.convert_path."""
-
-        for (src, tgt, func) in path:
-            hop = convert_hop(src, tgt, func, inspect.getmodule(func),
-                              inspect.getsource(func))
-            self.convert_path.append(hop)
-
-    def simulate_convert(self, tgt, src):
-        """Query networkx to see what the shortest path is from source to target.
-
-        Positional arguments:
-        tgt -- Destination object or type.
-        src -- Source object or type.
-
-        Calls self.add_convert to add the resulting path information to itself.
-
-        Raises networkx.NetworkXNoPath if no path from src to tgt exists."""
-
-        if not isinstance(tgt, type):
-            tgt = type(tgt)
-        if not isinstance(src, type):
-            src = type(src)
-
-        convert_path = convert.path(src, tgt)
-        self.add_convert(convert_path)
-
-    def __str__(self):
-        if isinstance(self.source, type):
-            source = self.source.__name__ + ' type'
-        else:
-            source = type(self.source).__name__ + ' object'
-        if isinstance(self.target, type):
-            target = self.target.__name__ + ' type'
-        else:
-            target = type(self.target).__name__ + ' object'
-
-        outstr = 'Odo path simulation from {src} to {tgt}\n'.format(
-            src=source, tgt=target)
-        outstr += '------------------------------------------------------------\n'
-        if self.append_sig:
-            outstr += 'Append will be called from {src} to {tgt}\n'.format(
-                src=self.append_sig[1].__name__,
-                tgt=self.append_sig[0].__name__)
-            outstr += 'Append will dispatch function {func} from {mod}\n'.format(
-                func=self.append_func.__name__, mod=self.append_mod.__name__)
-            outstr += 'Function {func} code follows:\n'.format(
-                func=self.append_func.__name__)
-
-            for line in self.append_code.split('\n'):
-                outstr += '    ' + line + '\n'
-
-        if self.convert_sig:
-            if isinstance(self.convert_sig[0], basestring):
-                target = self.convert_sig[0]
-            else:
-                target = self.convert_sig[0].__name__
-            outstr += 'Convert will be called from {src} to {tgt}\n'.format(
-                src=self.convert_sig[1].__name__, tgt=target)
-
-        if self.convert_path:
-            for (n, hop) in enumerate(self.convert_path, start=1):
-                outstr += 'Conversion hop {n}: {src} to {tgt}\n'.format(
-                    n=n, src=hop.source.__name__, tgt=hop.target.__name__)
-                outstr += 'Hop dispatches function {func} from {mod}\n'.format(
-                    func=hop.func.__name__, mod=hop.mod.__name__)
-                outstr += 'Function {func} code follows:\n'.format(
-                    func=hop.func.__name__)
-
-                for line in hop.code.split('\n'):
-                    outstr += '    ' + line + '\n'
-        elif self.convert_sig:
-           outstr += ('Convert pathing could not be determined (maybe {} is not '
-                    'a type in a reachable namespace?)'.format(self.convert_sig[0]))
-
-        return outstr
-
-    def __repr__(self):
-        return 'Path(source={src}, target={tgt})'.format(
-            src=repr(self.source), tgt=repr(self.target))
 
 @into.register(type, object)
 def into_type(a, b, dshape=None, simulate=False, **kwargs):
     if simulate:
-        path = Path(b, a)
-        path.convert_sig = (a, type(b))
-        path.simulate_convert(a, b)
-        return path
+        return Path(b, a, call_type='convert')
 
     with ignoring(NotImplementedError):
         if dshape is None:
@@ -233,30 +100,7 @@ def into_object(target, source, dshape=None, simulate=False, **kwargs):
     into.append.append      - Add things onto existing things
     """
     if simulate:
-        path = Path(source, target)
-        path.append_sig = (type(target), type(source))
-        path.add_append(append.dispatch(type(target), type(source)))
-
-        # Look for a convert call inside the append function code
-        match = re.search(r'convert\((.+?),', path.append_code, flags=re.I)
-        if match:
-            convert_str = match.group(1)
-            try:
-                # Convert object name string to actual object type, if possible
-                try:
-                    module = importlib.import_module('__builtin__')
-                    convert_type = getattr(module, convert_str)
-                except AttributeError:
-                    module, convert_str = convert_str.rsplit(".", 1)
-                    module = importlib.import_module(module)
-                    convert_type = getattr(module, convert_str)
-
-                path.convert_sig = (convert_type, type(source))
-                path.simulate_convert(convert_type, source)
-            except:
-                path.convert_sig = (convert_str, type(source))
-
-        return path
+        return Path(source, target, call_type='append')
 
     if isinstance(source, (str, unicode)):
         source = resource(source, dshape=dshape, **kwargs)
@@ -274,7 +118,6 @@ def into_string(uri, b, dshape=None, simulate=False, **kwargs):
     resource_ds = 0 * dshape.subshape[0] if isdimension(dshape[0]) else dshape
 
     a = resource(uri, dshape=resource_ds, expected_dshape=dshape, **kwargs)
-
     return into(a, b, dshape=dshape, simulate=simulate, **kwargs)
